Move StateAccessor, MergingStateAccessor to runners/core-java
