+++ runners/core-java/src/main/java/org/apache/beam/runners/core/MergingStateAccessor.java
@@ -15,12 +15,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.sdk.util.state;
+package org.apache.beam.runners.core;
 
 import java.util.Map;
 import org.apache.beam.sdk.annotations.Experimental;
 import org.apache.beam.sdk.annotations.Experimental.Kind;
 import org.apache.beam.sdk.transforms.windowing.BoundedWindow;
+import org.apache.beam.sdk.util.state.State;
+import org.apache.beam.sdk.util.state.StateTag;
 
 /**
  * Interface for accessing persistent state while windows are merging.
